requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
playwright==1.40.0
requests==2.31.0
groq==0.4.1
python-dotenv==1.0.0
aiofiles==23.2.1
pypdf==3.17.1
pandas==2.1.3
numpy==1.26.2
pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
